requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
openai==1.3.7
chromadb==0.4.18
python-dotenv==1.0.0
pypdf==3.17.4
flask==3.0.0 